+++ app/components/primer/primer.ts
@@ -8,6 +8,6 @@ import './alpha/tool_tip'
 import './alpha/x_banner'
 import './beta/auto_complete/auto_complete'
 import './beta/clipboard_copy'
-import './local_time'
+import './alpha/local_time'
 import './tab_container_component'
 import './time_ago_component'
