+++ static/statuses.json
@@ -20,6 +20,7 @@
   "Primer::Alpha::Layout": "alpha",
   "Primer::Alpha::Layout::Main": "alpha",
   "Primer::Alpha::Layout::Sidebar": "alpha",
+  "Primer::Alpha::LocalTime": "alpha",
   "Primer::Alpha::NavList": "alpha",
   "Primer::Alpha::NavList::Item": "alpha",
   "Primer::Alpha::NavList::Section": "alpha",
@@ -76,7 +77,7 @@
   "Primer::LabelComponent": "deprecated",
   "Primer::LayoutComponent": "alpha",
   "Primer::LinkComponent": "deprecated",
-  "Primer::LocalTime": "alpha",
+  "Primer::LocalTime": "deprecated",
   "Primer::Markdown": "beta",
   "Primer::MenuComponent": "alpha",
   "Primer::Navigation::TabComponent": "alpha",
