+++ test/components/component_test.rb
@@ -27,7 +27,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::Alpha::TabNav, { label: "label" }],
     [Primer::Alpha::UnderlinePanels, { label: "Panel label" }],
     [Primer::Alpha::Image, { src: "https://github.com/github.png", alt: "alt" }],
-    [Primer::LocalTime, { datetime: DateTime.parse("2014-06-01T13:05:07Z") }],
+    [Primer::Alpha::LocalTime, { datetime: DateTime.parse("2014-06-01T13:05:07Z") }],
     [Primer::Alpha::ImageCrop, { src: "Foo" }],
     [Primer::IconButton, { icon: :star, "aria-label": "Label" }],
     [Primer::Alpha::ActionList, { aria: { label: "Action List" } }, lambda do |component|
@@ -111,6 +111,7 @@ class PrimerComponentTest < Minitest::Test
 
   def test_registered_components
     ignored_components = [
+      "Primer::LocalTime",
       "Primer::ClipboardCopy",
       "Primer::LabelComponent",
       "Primer::LinkComponent",
